IndexedDB/tests/submissions/Opera/idbdatabase_deleteObjectStore4-not_reused.htm

<!DOCTYPE html>
<meta charset=utf-8>
<title>IDBDatabase.deleteObjectStore() - object store's name is removed from database's list </title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<script>

var t = async_test(),
    keys = [],
    open_rq = createdb(t)

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result

    var objStore = db.createObjectStore("resurrected", { autoIncrement: true, keyPath: "k" });
    objStore.add({k:5}).onsuccess = function(e) { keys.push(e.target.result); }
    objStore.add({}).onsuccess = function(e) { keys.push(e.target.result); }
    objStore.createIndex("idx", "i");
    assert_true(objStore.indexNames.contains("idx"));
    assert_equals(objStore.keyPath, "k", "keyPath");

    db.deleteObjectStore("resurrected");

    var objStore2 = db.createObjectStore("resurrected", { autoIncrement: true });
    objStore2.add("Unicorns'R'us").onsuccess = function(e) { keys.push(e.target.result); };
    assert_false(objStore2.indexNames.contains("idx"), "index exist on new objstore");
    assert_equals(objStore2.keyPath, null, "keyPath");

    assert_throws("NotFoundError", function() { objStore2.index("resurrected"); });
}

open_rq.onsuccess = function(e) {
    assert_object_equals(keys, [5, 6, 1], "keys");
    t.done();
}

</script>

<div id="log"></div>
